+++ lib/core/navigation/enums/route_type.dart
@@ -1,5 +1,1 @@
-enum FFRRouteType {
-  fullPage,
-  dialog,
-  bottomSheet,
-}
+enum FFRRouteType { fullPage, dialog, bottomSheet, action }