+++ lib/flutter_full_router.dart
@@ -1,6 +1,5 @@
 library;
 
-export 'core/navigation/actions/navigator_action.dart';
 export 'core/navigation/enums/open_flow.dart';
 export 'core/navigation/enums/route_type.dart';
 export 'core/navigation/models/route_definition.dart';