+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# CHANGELOG
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial open source release of FFR.
+- Added Support for basic route definitions.
+- Configurable Route Guard Interceptors.
+
+## 0.0.2
+
+- Action Routes via `FFRRouteType.action`.